refactor: change _cohere module to docindex_cohere
