+++ src/docindex_cohere/doc_index.py
